% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcleaning.R
\name{class_counts}
\alias{class_counts}
\title{Class Counts}
\usage{
class_counts(doc_tokens)
}
\arguments{
\item{doc_tokens}{the tokens in the document}
}
\value{
a named vector with total token count 'N' and unique token count 'V'
}
\description{
collects the length of total tokens (N),
and the number of unique tokens (V) in the collection of documents
}
\examples{
doc_tokens <- tokenize_docs( c('First entry of the doc,', 'Second entry of the doc.') )
class_counts(doc_tokens)
}
