% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{ClassifierCorpus}
\alias{ClassifierCorpus}
\title{Classifier Corpus}
\usage{
ClassifierCorpus(positive, negative)
}
\arguments{
\item{positive}{a document of text containing entries with positive sentiment}

\item{negative}{a document of text containing entries with negative sentiment}
}
\value{
an object of class ClassifierCorpus
}
\description{
a specific type of Corpus that includes samples from a 'positive' sentiment document
and a 'negative' sentiment document
}
\examples{
classifierCorpusEx <- ClassifierCorpus(c('what a beautiful day', 'I am happy'), c('what a miserable day', 'I am unhappy'))
}
