% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcleaning.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text}
\usage{
clean_text(text, stopwords = TRUE)
}
\arguments{
\item{text}{a line of text}

\item{stopwords}{logical - should stopwords be removed}
}
\value{
a cleaned line of text
}
\description{
performs basic text cleaning on a single line of text by
converting all text to lowercase and removing all punctuation and removing
stopwords using the stopwords selection maintained in the tm package
}
\examples{
clean_text('This is a basic line. It contains UNCLEANED text.')
}
