% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclassifier.R
\name{conditionalProbability}
\alias{conditionalProbability}
\title{Conditional Probabilities for Word by TermFrequency}
\usage{
conditionalProbability(token, termFrequency)
}
\arguments{
\item{token}{a single token}

\item{termFrequency}{an object of class TermFrequency}
}
\value{
a conditional probability for the token belonging to each class given it's term frequency
}
\description{
this function takes a token and calculates the conditional probability
of the token appearing in each classLabel.  This can then be used to get the likelihood ratio
of the token belonging to each classLabel.
}
\examples{
corpusEx <- Corpus(list('first' = c('a single entry document'), 'second' = c('a two entry document','with two separate entries')))
termFreqEx <- TermFrequency(corpusEx)
P_token <- conditionalProbability( 'entr', termFreqEx)
}
