% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\usage{
Corpus(documents, docnames = NULL)
}
\arguments{
\item{documents}{a list of one or more character vectors of text belonging to one or more documents}

\item{docnames}{the classnames for the different documents}
}
\value{
an object of class Corpus
}
\description{
a representation of a document or set of documents that includes the tokenized document entries
document counts, token counts, etc.
}
\examples{
corpusEx <- Corpus(list())
}
