% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{is.Corpus}
\alias{is.Corpus}
\title{Is Corpus}
\usage{
is.Corpus(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
true if the object inherits from Corpus
}
\description{
check if an object inherits from class Corpus
}
\examples{
x <- structure(class='Any')
is.Corpus(x) # -> FALSE
class(x) <- c(class(x), 'Corpus')
is.Corpus(x) # -> TRUE
}
