% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{length.Corpus}
\alias{length.Corpus}
\title{Length for Corpus}
\usage{
\method{length}{Corpus}(obj)
}
\arguments{
\item{obj}{an object inheriting from class Corpus}
}
\value{
the number of documents in the Corpus
}
\description{
length implementation for a Corpus object
returns the number of documents in the Corpus
}
