% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttransform.R
\name{minimum_edit_distance}
\alias{minimum_edit_distance}
\title{Minimum Edit Distance}
\usage{
minimum_edit_distance(
  source,
  target,
  deletecost = 1,
  insertcost = 1,
  replacecost = 2
)
}
\arguments{
\item{source}{\itemize{
\item the token starting point
}}

\item{target}{\itemize{
\item the token to compare against
}}

\item{deletecost}{\itemize{
\item the cost of deleting one letter from the target
}}

\item{insertcost}{\itemize{
\item the cost of inserting one letter into the target
}}

\item{replacecost}{\itemize{
\item the cost of replacing on letter from the target
}}
}
\value{
dist - the minimum edit distance between source and target, D - the filled in distance matrix steps ( for backtrace )
}
\description{
Finds the minimum cost (in terms of edit operations) to transform the source token to the target token
}
\examples{
minimum_edit_distance('play','stay') # -> dist = 4
}
