% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclassifier.R
\name{predict.SentimentClassifier}
\alias{predict.SentimentClassifier}
\title{Predict Sentiment Classifier}
\usage{
\method{predict}{SentimentClassifier}(obj, newdata)
}
\arguments{
\item{obj}{a SentimentClassifier model object}

\item{newdata}{a character vector containing text samples to be assessed}
}
\value{
Sentiment score predictions of each entry of newdata
}
\description{
a prediction method to assign a sentiment score to a set of text samples from a Sentiment Classifier
}
