% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{print.Corpus}
\alias{print.Corpus}
\title{Print For Corpus}
\usage{
\method{print}{Corpus}(obj)
}
\arguments{
\item{obj}{an object inheriting from class Corpus}
}
\value{
a printed representation of the Corpus
}
\description{
print implementation for a Corpus object
prints the class name and the number of tokens in each document
}
