% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclassifier.R
\name{SentimentClassifier}
\alias{SentimentClassifier}
\title{Sentiment Classifier}
\usage{
SentimentClassifier(positive, negative, type = c("linear", "bayes"))
}
\arguments{
\item{positive}{character vector with samples of positive text}

\item{negative}{character vector with samples of negative text}

\item{type}{the classifier model type (default \'linear\')}
}
\value{
a SentimentClassifier Object
}
\description{
a simple sentiment classifier object that takes a set of positive sample text
and a set of negative sample text, and creates a sentiment classifier.
}
\examples{
SentimentClassifier(c('some positive text'), c('negative text samples'))
}
