% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttermfrequency.R
\name{TermFrequency}
\alias{TermFrequency}
\title{Term Frequency}
\usage{
TermFrequency(corpus)
}
\arguments{
\item{corpus}{an object inheriting from class Corpus}
}
\value{
a term frequency hash for the tokens and documents in the corpus
}
\description{
a term frequency hash for the tokens and documents of the corpus
}
\examples{
corpusEx <- Corpus(list('first' = c('a single entry document'), 'second' = c('a two entry document','with two separate entries')))
termFreqEx <- TermFrequency(corpusEx)
}
