% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttransform.R
\name{token_frequency_check}
\alias{token_frequency_check}
\title{Token Frequency Check - DataMuse API}
\usage{
token_frequency_check(token)
}
\arguments{
\item{token}{\itemize{
\item a single token
}}
}
\value{
\itemize{
\item the token frequency per 1 million words in the English language, according to the DataMuse API
}
}
\description{
makes a call to the datamuse api to check the universal frequency of the token (count per million)
}
\examples{
token_frequency_check('the')
}
