% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttransform.R
\name{token_permutation.all}
\alias{token_permutation.all}
\title{Token Permutations - All}
\usage{
token_permutation.all(token, split = TRUE, simplify = FALSE)
}
\arguments{
\item{token}{\itemize{
\item either 1 token or a pre-split set of ordered letters that make up the token
}}

\item{split}{\itemize{
\item whether the input token needs to be split into individual letters or not - default = TRUE
}}

\item{simplify}{\itemize{
\item whether to return a simplified vector of the token permutations,
or a list containing the permutations for each separate operation (delete, replace, insert) - default = FALSE
}}
}
\value{
\itemize{
\item by default returns a list with the separate permutations for 1 delete, replace and insert
if simplify is true, returns a character vector with each possible permutation of the original token after 1 operation
}
}
\description{
Collects all tokens that could result from performing any edit operation on the input token using a more efficient single-pass implementation
}
\examples{
token_permutation.all('sometoken')
token_permutation.all(c('s','o','m','e','t','o','k','e','n'), split = FALSE)
token_permutation.all('sometoken', simplify = TRUE)
}
