% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttransform.R
\name{token_permutation.delete}
\alias{token_permutation.delete}
\title{Token Permutations - Delete}
\usage{
token_permutation.delete(token, split = TRUE)
}
\arguments{
\item{token}{\itemize{
\item either 1 token or a pre-split set of ordered letters that make up the token
}}

\item{split}{\itemize{
\item whether the input token needs to be split into individual letters or not - default = TRUE
}}
}
\value{
\itemize{
\item a character vector with each possible permutation of the original token after 1 delete operation
}
}
\description{
Collects all tokens that could result from performing 1 delete operation
on the given token
}
\examples{
token_permutations.delete('sometoken')
token_permutations.delete(c('s','o','m','e','t','o','k','e','n'), split = FALSE)
}
