% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttransform.R
\name{token_permutation.replace}
\alias{token_permutation.replace}
\title{Token Permutations - Replace}
\usage{
token_permutation.replace(token, split = TRUE)
}
\arguments{
\item{token}{\itemize{
\item either 1 token or a pre-split set of ordered letters that make up the token
}}

\item{split}{\itemize{
\item whether the input token needs to be split into individual letters or not - default = TRUE
}}
}
\value{
\itemize{
\item a character vector with each possible permutation of the original token after 1 replace operation
}
}
\description{
Collects all tokens that could result from replace 1 letter with a different letter anywhere in the given token
}
\examples{
token_permutation.replace('sometoken')
token_permutation.replace(c('s','o','m','e','t','o','k','e','n'), split = FALSE)
}
