% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcleaning.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Tokenize Text}
\usage{
tokenize_text(text, stopwords = TRUE)
}
\arguments{
\item{text}{a line of text}

\item{stopwords}{logical - should stopwords be removed}
}
\value{
a cleaned, stemmed and tokenized character vector
}
\description{
performs cleaning, word stemming, and tokenization on a single line of text
using clean_text and the stemming and boosting utilities maintained in the tm package
}
\examples{
tokenize_text('This is a basic line. It contains UNCLEANED, un-stemmed, un-tokenized text.')
}
