% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus.R
\name{unique.Corpus}
\alias{unique.Corpus}
\title{Uniquer for Corpus}
\usage{
\method{unique}{Corpus}(obj)
}
\arguments{
\item{obj}{an object inheriting from class Corpus}
}
\value{
a character vector containing each token in the corpus once
}
\description{
unique function implementation for a Corpus object
returns the vocabulary of unique tokens in the Corpus
}
