% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcleaning.R
\name{vocab_docs}
\alias{vocab_docs}
\title{Create Vocab for Docs}
\usage{
vocab_docs(docs)
}
\arguments{
\item{docs}{a vector or list of text samples}
}
\value{
a character vector containing each unique token entry that appears in the docs.
}
\description{
collect all unique tokens in the full set of docs which serve as the base vocabulary
for the corpus.  Returns a character vector of the unique tokens across all entries in the docs.
}
\examples{
vocab_docs( c('First entry of the doc,', 'Second entry of the doc.') )
}
